ai_backend/requirements.txt
torch
torchvision
torchaudio

numpy
pandas

matplotlib
seaborn

opencv-python
Pillow

scikit-learn

tqdm

albumentations

pyyaml
tqdm
duckduckgo-search